фиксим округление

+++ backend/models.go
@@ -22,44 +22,44 @@ type ReceiptResponse struct {
 type Item struct {
 	Name        string  `json:"name"`
 	NDS         int     `json:"nds"`
-	NDSSum      int64   `json:"ndsSum"` // Изменено на int64
+	NDSSum      float32 `json:"ndsSum"`
 	PaymentType int     `json:"paymentType"`
-	Price       int64   `json:"price"` // Изменено на int64
+	Price       float32 `json:"price"`
 	ProductType int     `json:"productType"`
 	ProviderINN string  `json:"providerInn"`
-	Quantity    float64 `json:"quantity"` // Оставлено как float64, так как количество может быть дробным
-	Sum         int64   `json:"sum"`      // Изменено на int64
+	Quantity    float32 `json:"quantity"`
+	Sum         float32 `json:"sum"`
 }
 
 type FiscalDataResponse struct {
-	BuyerPhoneOrAddress     string `json:"buyerAddress"`
-	CashTotalSum            int64  `json:"cashTotalSum"` // Изменено на int64
-	Code                    int    `json:"code"`
-	CreditSum               int64  `json:"creditSum"` // Изменено на int64
-	DateTime                string `json:"dateTime"`
-	ECashTotalSum           int64  `json:"ecashTotalSum"` // Изменено на int64
-	FiscalDocumentFormatVer string `json:"fiscalDocumentFormatVer"`
-	FiscalDocumentNumber    int    `json:"fiscalDocumentNumber"`
-	FiscalDriveNumber       string `json:"fiscalDriveNumber"`
-	FiscalSign              string `json:"fiscalSign"`
-	Items                   []Item `json:"items"`
-	KKTRegID                string `json:"kktRegId"`
-	NDS10                   int64  `json:"nds10"` // Изменено на int64
-	NDS18                   int64  `json:"nds18"` // Изменено на int64
-	OperationType           int    `json:"operationType"`
-	Operator                string `json:"operator"`
-	OperatorINN             string `json:"operatorInn"`
-	PrepaidSum              int64  `json:"prepaidSum"`   // Изменено на int64
-	ProvisionSum            int64  `json:"provisionSum"` // Изменено на int64
-	RequestNumber           int    `json:"requestNumber"`
-	RetailPlace             string `json:"retailPlace"`
-	RetailPlaceAddress      string `json:"retailPlaceAddress"`
-	ShiftNumber             int    `json:"shiftNumber"`
-	TaxationType            int    `json:"taxationType"`
-	AppliedTaxationType     int    `json:"appliedTaxationType"`
-	TotalSum                int64  `json:"totalSum"` // Изменено на int64
-	User                    string `json:"user"`
-	UserINN                 string `json:"userInn"`
+	BuyerPhoneOrAddress     string  `json:"buyerAddress"`
+	CashTotalSum            float32 `json:"cashTotalSum"`
+	Code                    int     `json:"code"`
+	CreditSum               float32 `json:"creditSum"`
+	DateTime                string  `json:"dateTime"`
+	ECashTotalSum           float32 `json:"ecashTotalSum"`
+	FiscalDocumentFormatVer string  `json:"fiscalDocumentFormatVer"`
+	FiscalDocumentNumber    int     `json:"fiscalDocumentNumber"`
+	FiscalDriveNumber       string  `json:"fiscalDriveNumber"`
+	FiscalSign              string  `json:"fiscalSign"`
+	Items                   []Item  `json:"items"`
+	KKTRegID                string  `json:"kktRegId"`
+	NDS10                   float32 `json:"nds10"`
+	NDS18                   float32 `json:"nds18"`
+	OperationType           int     `json:"operationType"`
+	Operator                string  `json:"operator"`
+	OperatorINN             string  `json:"operatorInn"`
+	PrepaidSum              float32 `json:"prepaidSum"`
+	ProvisionSum            float32 `json:"provisionSum"`
+	RequestNumber           int     `json:"requestNumber"`
+	RetailPlace             string  `json:"retailPlace"`
+	RetailPlaceAddress      string  `json:"retailPlaceAddress"`
+	ShiftNumber             int     `json:"shiftNumber"`
+	TaxationType            int     `json:"taxationType"`
+	AppliedTaxationType     int     `json:"appliedTaxationType"`
+	TotalSum                float32 `json:"totalSum"`
+	User                    string  `json:"user"`
+	UserINN                 string  `json:"userInn"`
 }
 
 type TransformedReceipt struct {
